Added current api routes in next frontend;added basic layout.
