+++ client/next.config.mjs
