--- client/src/App.css
@@ -1,61 +1,0 @@
-.app-container {
-  text-align: center;
-  font-family: Arial, sans-serif;
-  padding: 20px;
-}
-
-.title {
-  font-size: 3em;
-  margin: 20px 0;
-}
-
-.header {
-  font-size: 1.5em;
-  margin: 20px 0;
-}
-
-.input-container {
-  margin-top: 20px;
-}
-
-.input-box {
-  padding: 10px;
-  font-size: 1em;
-  width: 300px;
-  margin-right: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-}
-
-.add-button, .reset-button {
-  padding: 10px 20px;
-  font-size: 1em;
-  margin: 0 10px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.add-button {
-  background-color: #28a745;
-  color: white;
-}
-
-.reset-button {
-  background-color: #dc3545;
-  color: white;
-}
-
-.ingredient-list {
-  margin-top: 20px;
-}
-
-.ingredient-item {
-  margin: 5px 0;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  display: inline-block;
-  width: 200px;
-  text-align: left;
-}
