--- recipe-app/app/page.js
@@ -1,9 +1,0 @@
-import Image from "next/image";
-
-export default function Home() {
-  return (
-    <main>
-      <p> Welcome to the Recipe App!</p>
-    </main>
-  );
-}
